Analizo metric collector splitted as MetricFu was.
Missing unit tests for parser.
