+++ lib/metric_collector/native/analizo/parser.rb
@@ -1,0 +1,66 @@
+module MetricCollector
+  module Native
+    module Analizo
+      class Parser
+        attr_accessor :processing, :wanted_metrics
+
+        def initialize(processing: nil, wanted_metrics: [])
+          self.processing = processing
+          self.wanted_metrics = wanted_metrics
+        end
+
+        def parse_all(output)
+          YAML.load_documents(output).each do |hash|
+            parse(hash)
+          end
+        end
+
+        private
+
+        def remove_extension(name)
+          list = name.split(".")
+          if (list.size > 1)
+            list.delete_at(list.size - 1)
+          end
+          list.join(".")
+        end
+
+        def parse_file_name(file_name)
+          file_name_tokenized = file_name.to_s.split("/")
+          file_name_tokenized[file_name_tokenized.size - 1] = remove_extension(file_name_tokenized.last)
+          file_name_tokenized.join(".")
+        end
+
+        def new_metric_result(module_result, code, value)
+          MetricResult.create(metric: self.wanted_metrics[code].metric, value: value.to_f, module_result: module_result, metric_configuration_id: self.wanted_metrics[code].id)
+        end
+
+        def module_result(module_name, granularity)
+          kalibro_module = KalibroModule.new({long_name: module_name, granularity: granularity})
+          module_result = ModuleResult.find_by_module_and_processing(kalibro_module, self.processing)
+
+          if module_result.nil?
+            kalibro_module.save
+            ModuleResult.create(kalibro_module: kalibro_module, processing: self.processing)
+          else
+            module_result
+          end
+        end
+
+        def parse(result_map)
+          if result_map['_filename'].nil?
+            module_result = module_result("", Granularity::SOFTWARE)
+          else
+            module_result = module_result(parse_file_name(result_map['_filename'].last), Granularity::CLASS)
+          end
+
+          result_map.each do |code, value|
+            new_metric_result(module_result, code, value) if (self.wanted_metrics[code])
+          end
+
+          module_result
+        end
+      end
+    end
+  end
+end
