+++ spec/lib/metric_collector/base_spec.rb
@@ -11,5 +11,11 @@ describe MetricCollector::Base, :type => :model do
         expect { subject.collect_metrics("", "") }.to raise_error(NotImplementedError)
       end
     end
+
+    describe 'available?' do
+      it 'is expected to raise NotImplementedError' do
+        expect { MetricCollector::Base.available? }.to raise_error(NotImplementedError)
+      end
+    end
   end
 end
