--- spec/lib/metric_collector/native/analizo_spec.rb
@@ -1,69 +1,0 @@
-require 'rails_helper'
-require 'metric_collector'
-
-describe MetricCollector::Native::Analizo, :type => :model do
-  describe 'method' do
-    let(:analizo_metric_collector_list) { FactoryGirl.build(:analizo_metric_collector_list) }
-
-    describe 'available?' do
-      context 'when analizo is installed' do
-        before :each do
-          MetricCollector::Native::Analizo.expects(:`).with("analizo --version").returns(analizo_metric_collector_list.version)
-        end
-        it 'is expected to be truthy' do
-          expect(MetricCollector::Native::Analizo.available?).to be_truthy
-        end
-      end
-      context 'when analizo is not installed' do
-        before :each do
-          MetricCollector::Native::Analizo.expects(:`).with("analizo --version").returns(nil)
-        end
-        it 'is expected to be falsey' do
-          expect(MetricCollector::Native::Analizo.available?).to be_falsey
-        end
-      end
-    end
-
-    describe 'collect_metrics' do
-      include MockHelper
-      let!(:metric_collector_details) { FactoryGirl.build(:metric_collector_details) }
-      let!(:native_metric) { metric_collector_details.supported_metrics["acc"] }
-      let!(:wanted_metric_configuration) { FactoryGirl.build(:metric_configuration, metric: native_metric) }
-      let!(:absolute_path) { "app/models/metric.rb" }
-      let!(:module_result) { FactoryGirl.build(:module_result) }
-      let(:wanted_metrics_list) { [wanted_metric_configuration] }
-      let(:processing) { FactoryGirl.build(:processing) }
-
-      before :each do
-        subject.expects(:`).with("analizo metrics #{absolute_path}").returns(analizo_metric_collector_list.raw_result)
-        MetricCollector::Native::Analizo.any_instance.expects(:details).returns(metric_collector_details)
-        MetricResult.expects(:create).with(metric: native_metric,
-                                           value: analizo_metric_collector_list.parsed_result[1]["acc"].to_f,
-                                           module_result: module_result,
-                                           metric_configuration_id: wanted_metric_configuration.id).returns(FactoryGirl.build(:metric_result))
-      end
-
-      context 'when there are no ModuleResults with the same module and processing' do
-        before :each do
-          KalibroModule.any_instance.expects(:save).twice.returns(true)
-          ModuleResult.expects(:create).at_least_once.returns(module_result)
-          find_module_result_mocks
-        end
-
-        it 'should collect the metrics for a given project' do
-          expect(subject.collect_metrics(absolute_path, wanted_metrics_list, processing)).to eq(analizo_metric_collector_list.parsed_result)
-        end
-      end
-
-      context 'when there is an existing ModuleResult with the same module and processing' do
-        before :each do
-          find_module_result_mocks([module_result])
-        end
-
-        it 'should collect the metrics for a given project' do
-          expect(subject.collect_metrics(absolute_path, wanted_metrics_list, processing)).to eq(analizo_metric_collector_list.parsed_result)
-        end
-      end
-    end
-  end
-end
